extract mock class out and update copyright
